-support multi vpu instances between other applications    add shared memory oal, attach buffer class, ... -add encoder/muxer C API -support FFMpeg mp4,mov Muxer -add video encoder test application 'apps/multimedia/mme_enc' -test encoder(H263) decoder(VP8, DIV3, RV40)

+++ libs/libmme/MmpDecoder/MmpDecoderAudio_Ffmpeg.cpp
@@ -293,7 +293,7 @@ MMP_RESULT CMmpDecoderAudio_Ffmpeg::DecodeAu(CMmpMediaSample* pMediaSample, CMmp
     avcodec_get_frame_defaults(m_pAVFrame_Decoded);
     usebyte = avcodec_decode_audio4(m_pAVCodecContext, m_pAVFrame_Decoded, &frameFinished, &avpkt);
     if(usebyte > 0) {
-        if(usebyte > pMediaSample->uiAuSize) {
+        if(usebyte > (int32_t)pMediaSample->uiAuSize) {
             pDecResult->uiAuUsedByte = pMediaSample->uiAuSize;
         }
         else {
